@@ -6,45 +6,45 @@
     "start": "node_modules/.bin/ng serve",
     "build": "node_modules/.bin/ng build",
     "build:dev": "node_modules/.bin/ng build --aot=false",
-    "test": "node_modules/.bin/ng test --watch=false --codeCoverage=true --browsers=ChromeHeadless",
+    "test": "node_modules/.bin/ng test --no-watch --code-coverage --browsers=ChromeHeadless",
     "test:dev": "node_modules/.bin/ng test --watch --codeCoverage=true --progress --browsers=ChromeHeadless",
     "lint": "node_modules/.bin/npx ng lint exemple-ui",
     "e2e": "node_modules/.bin/ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.11",
-    "@angular/common": "^13.3.11",
-    "@angular/compiler": "^13.3.11",
-    "@angular/core": "^13.3.11",
-    "@angular/forms": "^13.3.11",
-    "@angular/localize": "^13.3.11",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
+    "@angular/animations": "^14.2.8",
+    "@angular/common": "^14.2.8",
+    "@angular/compiler": "^14.2.8",
+    "@angular/core": "^14.2.8",
+    "@angular/forms": "^14.2.8",
+    "@angular/localize": "^14.2.8",
+    "@angular/platform-browser": "^14.2.8",
+    "@angular/platform-browser-dynamic": "^14.2.8",
+    "@angular/router": "^14.2.8",
     "@ngxs/router-plugin": "^3.7.2",
     "@ngxs/storage-plugin": "^3.7.2",
     "@ngxs/store": "^3.7.2",
     "fast-json-patch": "^3.0.0-1",
     "moment": "^2.29.0",
     "primeflex": "~2.0.0",
-    "primeicons": "^2.0.0",
-    "primeng": "~12.0.0",
+    "primeicons": "^6.0.1",
+    "primeng": "~14.1.2",
     "rxjs": "~6.6.3",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^13.1.0",
-    "@angular-devkit/build-angular": "^13.3.7",
-    "@angular-eslint/builder": "13.2.1",
-    "@angular-eslint/eslint-plugin": "13.2.1",
-    "@angular-eslint/eslint-plugin-template": "13.2.1",
-    "@angular-eslint/schematics": "13.2.1",
-    "@angular-eslint/template-parser": "13.2.1",
-    "@angular/cli": "^13.3.7",
-    "@angular/compiler-cli": "^13.3.11",
-    "@angular/language-service": "^13.3.11",
+    "@angular-builders/custom-webpack": "^14.0.1",
+    "@angular-devkit/build-angular": "^14.2.7",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.7",
+    "@angular/compiler-cli": "^14.2.8",
+    "@angular/language-service": "^14.2.8",
     "@cucumber/cucumber": "^7.2.1",
     "@sinonjs/fake-timers": "^9.1.2",
     "@types/chai": "^4.1.7",
@@ -67,6 +67,7 @@
     "protractor-cucumber-framework": "^8.1.1",
     "sinon": "^11.1.1",
     "typescript": "~4.6.4",
-    "util": "^0.12.4"
+    "util": "^0.12.4",
+    "webpack": "^5.74.0"
   }
 }
